requirements.txt
feedparser
requests
python-telegram-bot